+++ src/models/strategy/sma-model.ts
@@ -1,6 +1,0 @@
-import { StrategySettingsModel } from "./strategy-settings-model";
-
-export interface SMAModel extends StrategySettingsModel {
-  shortPeriod: number;
-  longPeriod: number;
-}