+++ src/models/strategy/sma-settings-model.ts
@@ -1,0 +1,4 @@
+export interface SMASettingsModel {
+    shortPeriod: number;
+    longPeriod: number;
+}